[ALLUXIO-2913] Separate LineageMaster.java into an interface class and an implementation class
